fix(platform): 32-bit compile safety + ledger reconciliation (P4-QUAL-04)
CodeRabbit's third review pass found the pid upper-bound check
(pid > math.MaxUint32) would overflow the untyped constant on
windows/386 (32-bit int) and fail to compile. Widened to
uint64(pid) > math.MaxUint32, safe because pid <= 0 already
short-circuits before the conversion runs.

Also reconciles docs/audits/README.md: P4-QUAL-04's own ledger row
already scoped itself to "the Windows build only" as of 2026-07-05,
and this PR ships exactly that, so it moves to Recently shipped.
Closes out P4-GIT-07 (PR #197) the same way, per that PR's own
deferred-to-orchestrator note.

+++ internal/platform/procalive_windows.go
@@ -26,7 +26,11 @@ import (
 // requested to minimize the chance OpenProcess denies access, since a
 // combined access mask can only succeed if every requested right is granted.
 func ProcessAlive(pid int) bool {
-	if pid <= 0 || pid > math.MaxUint32 {
+	// Widen to uint64 before comparing: on windows/386 (32-bit int), `pid >
+	// math.MaxUint32` would overflow the untyped constant into int and fail
+	// to compile. DevStrap doesn't ship a 386 build today, but this file must
+	// still compile for any windows/GOARCH the module supports.
+	if pid <= 0 || uint64(pid) > math.MaxUint32 {
 		return false
 	}
 	handle, err := windows.OpenProcess(windows.SYNCHRONIZE, false, uint32(pid))
